pyproject.toml
[tool.poetry]
name = "bgbench"
version = "0.1.0"
description = "A benchmarking tool"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "bgbench"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
openai = "^1.58.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.9"
python-chess = "^1.999"
pokerkit = "^0.5.4"
litellm = "^1.60.0"
ruff = "^0.9.7"
numpy = {version = "^2.2.3", python = ">=3.11,<3.14"}
pymc = {version = "^5.21.1", python = ">=3.11,<3.14"}
arviz = {version = "^0.21.0", python = ">=3.11,<3.14"}
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.391"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
